requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
defusedxml==0.8.0rc2
distlib==0.3.8
Django==5.1
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
filelock==3.15.4
idna==3.7
mysqlclient==2.1.1
oauthlib==3.2.2
pipenv==2024.0.1
platformdirs==4.2.2
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==72.1.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.4.3
urllib3==2.2.2
virtualenv==20.26.3
